refactor: Move LuisRecognizerProxy to Common
